requirements.txt
# AutoCaptions - Required Python packages
openai-whisper
